// projects/maestro-ng/src/lib/models/consulta/restricao-operador.enum.ts
export enum RestricaoOperador {
  START_WITH = 'START_WITH',
  NOT_START_WITH = 'NOT_START_WITH',
  END_WITH = 'END_WITH',
  NOT_END_WITH = 'NOT_END_WITH',
  CONTAINS = 'CONTAINS',
  NOT_CONTAINS = 'NOT_CONTAINS',
  IN = 'IN',
  NOT_IN = 'NOT_IN',
  IS_NULL = 'IS_NULL',
  IS_NOT_NULL = 'IS_NOT_NULL',
  BETWEEN = 'BETWEEN',
  NOT_BETWEEN = 'NOT_BETWEEN',
  EQUALS = 'EQUALS',
  NOT_EQUALS = 'NOT_EQUALS',
  GT = 'GT',
  GE = 'GE',
  LT = 'LT',
  LE = 'LE',
}